+++ scripts/tag_monkey.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-import sys
-import common.tag
-
-common.tag.tag(
-    'src/monkeymaster/monkeymaster.js', 
-    prefix='monkeymaster')